+++ .my_cmds.sh
@@ -1,7 +1,7 @@
 # Add below line to .bashsrc and/or .bash_profile
 # source ~/.my_cmds.sh
-alias langs="cd 'C:\Users\xiang\Desktop\repos\useful-scripts\Langs' && python lang_selector.py"
-alias themes="cd 'C:\Users\xiang\Desktop\repos\useful-scripts\Themes' && python change_theme.py"
-alias recs="cd 'C:\Users\xiang\Desktop\repos\useful-scripts\Recs' && python recs.py"
-alias exercise="cd 'C:\Users\xiang\Desktop\repos\useful-scripts\Exercise' && pipenv run python exercise.py"
-alias aaa="cd 'C:\Users\xiang\Desktop\repos\useful-scripts\AAA' && pipenv run python aaa.py"
+alias langs="cd 'C:\Users\xiang\Desktop\Codebase\useful-scripts\Langs' && python lang_selector.py"
+alias themes="cd 'C:\Users\xiang\Desktop\Codebase\useful-scripts\Themes' && python change_theme.py"
+alias recs="cd 'C:\Users\xiang\Desktop\Codebase\useful-scripts\Recs' && python recs.py"
+alias exercise="cd 'C:\Users\xiang\Desktop\Codebase\useful-scripts\Exercise' && pipenv run python exercise.py"
+alias aaa="cd 'C:\Users\xiang\Desktop\Codebase\useful-scripts\AAA' && pipenv run python aaa.py"